// gui-game/src/app/constants/enums.ts
export enum GAME_MODE {
    PvC = "player vs player",
    CvC = "computer vs computer"
}

export enum GAME_TYPE {
    RPS = "rps",
    RPSLS = "rpls"
}

export enum GAME_STATE {
    WIN = "win",
    LOSE = "loose",
    TIE = "draw"
}

export enum EXCEPTIONS {
    INVALID_INPUT = "invalid input"
}

export enum PLAYER_STATE {
    START_GAME = 1,
    RETRY = 2,
    PLAY_AGAIN = 3
}

export enum PLAYERS {
    PLAYER_1,
    PLAYER_2
}